got validation working correctly
