+++ Bangazon/Controllers/PaymentTypesController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Bangazon.Models.ViewModels;
 
 namespace Bangazon.Controllers
 {
@@ -49,11 +50,17 @@ namespace Bangazon.Controllers
         // POST: PaymentTypes/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create(PaymentType paymentType)
+        public async Task<ActionResult> Create(PaymentTypeViewModel paymentTypeViewModel)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                return View(paymentType);
+                return View(new PaymentType 
+                {
+                    DateCreated = DateTime.Now,
+                    Description = paymentTypeViewModel.Description,
+                    AccountNumber = paymentTypeViewModel.AccountNumber,
+                    ExpirationDate = paymentTypeViewModel.ExpirationDate
+                });
             }
 
             try
@@ -62,19 +69,17 @@ namespace Bangazon.Controllers
 
                 var paymentTypeInstance = new PaymentType
                 {
-                    Description = paymentType.Description,
-                    AccountNumber = paymentType.AccountNumber,
+                    DateCreated = DateTime.Now,
+                    Description = paymentTypeViewModel.Description,
+                    AccountNumber = paymentTypeViewModel.AccountNumber,
+                    ExpirationDate = paymentTypeViewModel.ExpirationDate
                 };
-                    if (paymentType.ExpirationDate < DateTime.Now)
-                    {
-                        paymentTypeInstance.ExpirationDate = paymentType.ExpirationDate;
-                    }
 
                 paymentTypeInstance.UserId = user.Id;
 
-
                 _context.PaymentType.Add(paymentTypeInstance);
                 await _context.SaveChangesAsync();
+
 
                 return RedirectToAction(nameof(Index));
             }
